requirements.txt
# Generated by update_essential_files.py
# Date: 2025-04-08 17:31:15
# Updated for Python 3.10 compatibility

black==21.9b0
dependency_injector==4.40.0
fastapi==0.68.0
flake8==3.9.2
hypothesis==6.0.0
isort==5.9.3
matplotlib==3.4.0
modelcontextprotocol==0.1.0
mypy==0.910
networkx==2.5
numpy>=1.22.0  # Updated for Python 3.10 compatibility
pandas>=1.4.0  # Updated for compatibility with newer numpy
pre-commit==2.15.0
pydantic==1.8.2
pylint==2.11.1
pytest==6.2.5
pyyaml==6.0
rdflib==6.0.0
requests==2.25.0
sqlalchemy==2.0.23
alembic==1.12.0
psycopg2-binary==2.9.9
uvicorn==0.15.0
